No `s-o-b` when use `git-together commit`
By default, the `git-together commit` will add `Signed-off-by:` to the
committer. This will be redundant after `git author` setup the commit
template. Hence, we should disable this behavior.

In this case, the `git together commit` can still alternate the author
and committer, but it will use the template populated by `git author` to
capture all the authors.